Removed all aspects of multiple rounds
Was distracting from main purpose of this project and something I can add when I have more experience. Was creating some bugs and I was unable to keep track of score. Also the user doesn't have the ability to call on commands (game.restart, game.score) which was part of the reason I was implementing to get some experience with OOP and classes.